web/requirements.txt
fastapi
uvicorn
psutil
jinja2
pymongo
python-dotenv